requirements.txt
fastapi
uvicorn[standard]
pytesseract
pillow
numpy
python-multipart
opencv-python-headless
pyspellchecker